# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Velorki API
  version: 1.0.0
  summary: Relay service for the Velorki bike app.
  description: |
    The Velorki app is open source and therefore cannot hold client secrets.
    This service keeps them: it relays OAuth token exchanges to Strava and
    Ride with GPS, verifies entitlements against RevenueCat, proxies route
    planning to a language model, and hosts share links.

    All errors share one body shape:

    ```json
    { "error": { "code": "invalid_request", "message": "...", "retry_after_s": 30 } }
    ```

    Every response carries `X-Request-Id`, echoing the request header when one
    was supplied and generated otherwise. Clients should send
    `X-Velorki-Client` (e.g. `android/1.0.0+1`); it is logged when present.

    Rate limits are per-process in-memory token buckets. A rejected request
    answers 429 with `retry_after_s` in the body and a `Retry-After` header.
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
servers:
  - url: https://api.velorki.app
    description: Production
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /oauth/strava/token:
    post:
      tags: [oauth]
      operationId: stravaToken
      summary: Exchange a Strava authorization code for tokens.
      description: |
        Adds `client_id` and `client_secret` and posts the form-encoded
        exchange to `https://www.strava.com/oauth/token` with
        `grant_type=authorization_code`. Strava's JSON is returned unchanged,
        with its status code, on success.

        `redirect_uri` must match one entry of `OAUTH_REDIRECT_ALLOWLIST`
        exactly; prefixes and wildcards are not accepted.

        Rate limit: 10 per minute and 60 per day, per IP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthCodeRequest'
      responses:
        '200':
          $ref: '#/components/responses/OAuthTokenPassthrough'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/NotEntitled'
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          $ref: '#/components/responses/UpstreamError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /oauth/strava/refresh:
    post:
      tags: [oauth]
      operationId: stravaRefresh
      summary: Refresh a Strava access token.
      description: |
        Posts `grant_type=refresh_token` to Strava and returns the answer
        unchanged. Rate limit: 30 per hour, per IP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthRefreshRequest'
      responses:
        '200':
          $ref: '#/components/responses/OAuthTokenPassthrough'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/NotEntitled'
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          $ref: '#/components/responses/UpstreamError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /oauth/rwgps/token:
    post:
      tags: [oauth]
      operationId: rwgpsToken
      summary: Exchange a Ride with GPS authorization code for a token.
      description: |
        Posts `client_id`, `client_secret`, `code`, `redirect_uri` and
        `grant_type=authorization_code` to
        `https://ridewithgps.com/oauth/token.json` and returns the answer
        unchanged. `redirect_uri` must be on `OAUTH_REDIRECT_ALLOWLIST`.

        Rate limit: 10 per minute and 60 per day, per IP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthCodeRequest'
      responses:
        '200':
          $ref: '#/components/responses/OAuthTokenPassthrough'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/NotEntitled'
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          $ref: '#/components/responses/UpstreamError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /oauth/rwgps/refresh:
    post:
      tags: [oauth]
      operationId: rwgpsRefresh
      summary: Not implemented.
      description: |
        Ride with GPS access tokens do not expire, so there is nothing to
        refresh. Always answers 501 with the error code `unavailable`; the
        route exists so the app receives a structured answer rather than a 404.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthRefreshRequest'
      responses:
        '401':
          $ref: '#/components/responses/NotEntitled'
        '501':
          description: Not implemented.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                notImplemented:
                  value:
                    error:
                      code: unavailable
                      message: >-
                        Ride with GPS does not support refresh tokens;
                        re-authorize instead.
...
  /share:
    post:
      tags: [share]
      operationId: createShare
      summary: Store a route or ride and return a public link.
      description: |
        The GPX is stored verbatim, together with a small summary, under a
        random 10-character base62 id. Shares expire after 365 days and a daily
        sweep removes expired rows. Nothing links the row to the rider who
        created it.

        Rate limit: 30 per day, per app user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShareRequest'
      responses:
        '201':
          description: The share was created.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
                required: [id, url]
                additionalProperties: false
                properties:
                  id:
                    type: string
                    pattern: '^[A-Za-z0-9]{10}$'
                  url:
                    type: string
                    format: uri
              examples:
                created:
                  value:
                    id: 7Kq2mZ0aTb
                    url: https://api.velorki.app/s/7Kq2mZ0aTb
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/NotEntitled'
        '429':
          $ref: '#/components/responses/RateLimited'
  /s/{id}:
    get:
      tags: [share]
      operationId: getSharePage
      summary: The public share page.
      description: |
        Unauthenticated. A self-contained HTML page: inline CSS and JS, a
        pinned MapLibre GL build and the OpenFreeMap "liberty" style as the
        only external resources. The GPX is parsed in the browser. Unknown,
        malformed and expired ids all return 404.
      security: []
      parameters:
        - $ref: '#/components/parameters/ShareId'
      responses:
        '200':
          description: The share page.
          content:
            text/html:
              schema:
                type: string
        '404':
          $ref: '#/components/responses/NotFound'
  /s/{id}.gpx:
    get:
      tags: [share]
      operationId: getShareGpx
      summary: The shared GPX file.
      description: Unauthenticated. Returns the stored GPX byte for byte.
      security: []
      parameters:
        - $ref: '#/components/parameters/ShareId'
      responses:
        '200':
          description: The GPX document.
          content:
            application/gpx+xml:
              schema:
                type: string
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    revenueCatAppUserId:
      type: http
      scheme: bearer
      bearerFormat: revenuecat_app_user_id
      description: |
        `Authorization: Bearer <revenuecat_app_user_id>`.

        The service verifies the id against
        `GET https://api.revenuecat.com/v1/subscribers/{app_user_id}` using its
        own RevenueCat secret key. The caller is entitled when the configured
        entitlement (`REVENUECAT_ENTITLEMENT`, default `plus`) has an
        `expires_date` that is null or in the future. Answers are cached for
        10 minutes (entitled) or 1 minute (not entitled).

        When `REVENUECAT_MODE=stub` every caller is entitled; that mode is for
        forks and local development. A missing or malformed header is 401
        `not_entitled`.
  headers:
    XRequestId:
      description: Echo of the request's X-Request-Id, or a generated one.
      schema:
        type: string
    RetryAfter:
      description: Seconds to wait before retrying.
      schema:
        type: integer
        minimum: 1
  parameters:
    ShareId:
      name: id
      in: path
      required: true
      description: The 10-character base62 share id.
      schema:
        type: string
        pattern: '^[A-Za-z0-9]{10}$'
...
  schemas:
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              enum:
                - invalid_request
                - not_entitled
                - consent_required
                - not_found
                - rate_limited
                - upstream_error
                - unavailable
            message:
              type: string
            retry_after_s:
              type: number
              minimum: 0
              description: Present on 429 responses.
    OAuthCodeRequest:
      type: object
      required: [code, redirect_uri]
      additionalProperties: false
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 512
        redirect_uri:
          type: string
          minLength: 1
          maxLength: 2048
          description: Must match an entry of OAUTH_REDIRECT_ALLOWLIST exactly.
    OAuthRefreshRequest:
      type: object
      required: [refresh_token]
      additionalProperties: false
      properties:
        refresh_token:
          type: string
          minLength: 1
          maxLength: 2048
...
    ShareRequest:
      type: object
      required: [kind, name, gpx, summary]
      additionalProperties: false
      properties:
        kind:
          type: string
          enum: [route, ride]
        name:
          type: string
          minLength: 1
          maxLength: 200
        gpx:
          type: string
          minLength: 1
          maxLength: 2097152
          description: A GPX 1.1 document, at most 2 MB.
        summary:
          type: object
          required: [distance_km]
          additionalProperties: false
          properties:
            distance_km: { type: number, minimum: 0 }
            ascent_m: { type: number, minimum: 0 }
            duration_s: { type: number, minimum: 0 }
  responses:
    OAuthTokenPassthrough:
      description: |
        The provider's JSON response, unchanged, with its own status code.
        The shape is defined by Strava or Ride with GPS, not by this service.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true
          examples:
            strava:
              value:
                token_type: Bearer
                expires_at: 1893456000
                expires_in: 21600
                refresh_token: '...'
                access_token: '...'
                athlete: { id: 7 }
    InvalidRequest:
      description: The request body or parameters did not validate.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            badRedirect:
              value:
                error:
                  code: invalid_request
                  message: redirect_uri is not allowed.
    NotEntitled:
      description: Missing, malformed or unentitled Authorization header.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            notEntitled:
              value:
                error:
                  code: not_entitled
                  message: An active Velorki subscription is required.
...
    NotFound:
      description: Unknown, malformed or expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            notFound:
              value:
                error:
                  code: not_found
                  message: No such share.
    RateLimited:
      description: A rate limit was exceeded.
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            limited:
              value:
                error:
                  code: rate_limited
                  message: Too many requests. Please slow down.
                  retry_after_s: 6
    UpstreamError:
      description: |
        The upstream provider failed or answered with a non-2xx status. The
        provider's message is forwarded, scrubbed of any client secret.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            upstream:
              value:
                error:
                  code: upstream_error
                  message: Bad Request
    Unavailable:
      description: |
        The integration this route needs is not configured on this server, or a
        global budget is exhausted.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            unconfigured:
              value:
                error:
                  code: unavailable
                  message: Strava is not configured on this server.
security:
  - revenueCatAppUserId: []